Allow for multiple types of player status querying plugins
My stats program logs when players are online or offline. With this change, I can make another plugin with "isMySQLplugin" as false, enable it, and use it to query for player status instead of using Minecraft query.

This also fixes the player status on the highscores page. It was using the html name instead of String username, everyone would always be offline.