Add functionality to have stats only show up on certain pages
Currently those pages are "player", "highscores", and "grandTotals". Setting a stats/groups "display" config value to an array of any of these values will have that stat/group only show up on those pages. Setting it to the normal TRUE or FALSE will include it on all/no pages, like normal. Groups will still display a stat total even if the stat itself is not set to be shown on the page (also like normal).